FIX: move some more methods between classes
